Add rotationYawPitchRollNeg and additional tests for rotationYawPitchRoll
rotationYawPitchRollNeg isn't used but I added as a learning experience.
These may not be correct but the operations seem to be working.
Specifically the first tests rotate the individual axis +10 deg and then
-10 deg and the result goes back to the zero matrix or the "Unit" matrix
with 1's on the diagonal.

Another test I apply all rotation at once and then undo the the rotations
by applying the neg values one axis at a time in the reverse order applied
by rotationYawPitchRoll.

And Finally I create rotationYawPitchRollNeg which returns a matrix that
applied the rotation in the reverse order, which you can then use that
single matrix to undo the result of the rotationYawPitchRoll. Again you
end up with the "zero" rotation matrix.