Create math3dx.zig from math3d.zig
Right now window.render isn't working, the cube isn't a cube it's "plate".
So I've created a simpler test which is just a point at the center of the
window and then "rotating" it 90deg's. The point should just stay in the
middle, but it moves which is wrong.

I'm going to need to change things so I've created math3dx from math3d
and reorganized some things so I can make changes but not loose what I've
got.