Clean up window debug output
Rename window.render test to window.render.cube

Add window.pts test which uses input interface to make testing
interactive and easier.

Note: Still not working the cube isn't a "cube" and window.pts "line"
formed by the two points is on wrong side of the center point.
And if z of the the "line" is zero its off the screen. So something is
still definitely wrong with the code and my understanding :)
the two