Change window.keyctrl.trinagle to be asymmetrical
With this change I see that positive x-axis is to the left
instead of the right. Both Y and Z seem to be correct, Y is up
and Z is toward the camera directly out of the screen pointing
at the viewer.

Note to self, in the future use simple asymmetrical objects like
this triangle instead of symmetrical objects like a cube.