changed showImage to showTexture
We see "ypinkypink" in the texture. Not quite right, but there is
something. But when we try to use the texture on the plane we see
just black.

But if we set entity.texture to null we do see the plane so I think I
know the problem, we haven't created a uv's for the plane.